wip: fixing tests
