+++ src/Kafka/CreateKafkaInstanceWithSizes/components/InstanceInfo.test.tsx
@@ -1,4 +1,3 @@
-import { userEvent } from "@storybook/testing-library";
 import { composeStories } from "@storybook/testing-react";
 import { render, waitForI18n } from "../../../test-utils";
 import * as stories from "./InstanceInfo.stories";
@@ -15,20 +14,5 @@ describe("InstanceInfo", function () {
     const tree = await render(<Story />);
     await waitForI18n(tree);
     expect(tree.baseElement).toMatchSnapshot();
-  });
-
-  it("should show only a skeleton and a link to the quick start guides when loading", async function () {
-    const spy = jest.fn();
-    const { Loading } = composedStories;
-    const comp = render(<Loading onClickQuickStart={spy} />);
-    await waitForI18n(comp);
-    const instanceInfo = await comp.findByTestId("instance-info");
-    expect(instanceInfo.textContent).toBe(
-      "DetailsLoading contentsNeed help getting started?Follow our quick start guide."
-    );
-
-    userEvent.click(await comp.findByText("Follow our quick start guide."));
-
-    expect(spy).toBeCalledTimes(1);
   });
 });
